requirements.txt
pandas~=2.2.3
aiosqlite~=0.21.0
aiogram~=3.18.0
python-dotenv~=1.0.1